Add a button to _exclude_ terms from search
When digging through logs, I find myself clicking the "Add to current search"
button and then prefixing the term with a "-" to exclude it to filter out noisy
stuff (such as Zookeeper logs, Datadog warnings, etc). It'd be way faster to
just have a button to do that in one click.